Use `#[component]` macro for core components, deleting `leptos_core` package
