+++ Cargo.toml
@@ -3,7 +3,6 @@ members = [
   # core
   "leptos",
   "leptos_dom",
-  "leptos_core",
   "leptos_macro",
   "leptos_reactive",
   "leptos_server",
