--- leptos_core/Cargo.toml
@@ -1,46 +1,0 @@
-[package]
-name = "leptos_core"
-version = "0.0.18"
-edition = "2021"
-authors = ["Greg Johnston"]
-license = "MIT"
-repository = "https://github.com/gbj/leptos"
-description = "Core functionality for the Leptos web framework."
-
-[dependencies]
-leptos_dom = { path = "../leptos_dom", default-features = false, version = "0.0.18" }
-leptos_macro = { path = "../leptos_macro", default-features = false, version = "0.0.18" }
-leptos_reactive = { path = "../leptos_reactive", default-features = false, version = "0.0.18" }
-log = "0.4"
-typed-builder = "0.11"
-
-[dev-dependencies]
-leptos = { path = "../leptos", default-features = false, version = "0.0" }
-
-[features]
-csr = [
-	"leptos/csr",
-	"leptos_dom/web",
-	"leptos_macro/csr",
-	"leptos_reactive/csr",
-]
-hydrate = [
-	"leptos/hydrate",
-	"leptos_dom/web",
-	"leptos_macro/hydrate",
-	"leptos_reactive/hydrate",
-]
-ssr = [
-	"leptos/ssr",
-	"leptos_macro/ssr",
-	"leptos_reactive/ssr",
-]
-stable = [
-	"leptos/stable",
-	"leptos_dom/stable",
-	"leptos_macro/stable",
-	"leptos_reactive/stable",
-]
-
-[package.metadata.cargo-all-features]
-denylist = ["stable"]
