--- leptos_core/src/lib.rs
@@ -1,12 +1,0 @@
-#![deny(missing_docs)]
-
-//! This crate contains several utility pieces that depend on multiple crates.
-//! They are all re-exported in the main `leptos` crate.
-
-mod suspense;
-mod transition;
-
-pub use suspense::*;
-pub use transition::*;
-
-pub use typed_builder::TypedBuilder;
